+++ Furniture_management/Product.h
@@ -1,9 +1,11 @@
+//Product.h by Group BITI-07
+//Group memeber: Chee Fang Yee, Soon Boon Ming, Muhd Aiman Hakimi Bin Mohd Hafizi, Wong Man Yee, Lim Jia Hui
 #ifndef PRODUCT_H
 #define PRODUCT_H
 #include <string>
 using namespace std;
 
-//Think this class Product as Node in linked list.
+//Think this class Product as Node in linked list. By Chee Fang Yee.
 class Product {
 public:
 	string PID;
@@ -13,7 +15,7 @@ public:
 	int stock;
 
 	//each node has info like id, name, colour...
-	Product(string id, string productName, string colour, double price, int stock);
+	Product(string id, string productName, string colour, float price, int stock);
 	Product* next;
 };
